+++ R/Bn.R
@@ -34,7 +34,7 @@ bn <- function(group_id, md = NULL, data = NULL) {
         md <- as.matrix(dist(data) ^ 2)
     }
 
-    if (class(md) != "matrix") {
+    if (sum("matrix" %in% class(md)) == 0) {
         stop("md is not of class matrix")
     }