+++ NEWS.md
@@ -1,3 +1,8 @@
 # NEWS
 
-* First public release
+## uclust 0.2.0
+
+* Fixed small bug in uhclust and uclust (no longer returns random error)
+* Compatibility with R version 4.0 (class(matrix) returns vector)
+* export rep_optimBn function 
+